introduce matchmaking pool structure

+++ internal/matchmaking/rbt.go
@@ -2,7 +2,7 @@ package matchmaking
 
 type nodeKey struct {
 	playerId string
-	rating   int
+	rating   float64
 }
 
 // redBlackNode represents the Red-Black Tree node.
@@ -101,7 +101,7 @@ func (t *redBlackTree) removeNode(z *redBlackNode) {
 	t.size--
 }
 
-func (t *redBlackTree) search(rating int, playerId string) *redBlackNode {
+func (t *redBlackTree) search(rating float64, playerId string) *redBlackNode {
 	x := t.root
 
 	for x != t.leaf {
@@ -369,7 +369,7 @@ func (t *redBlackTree) transplant(u, v *redBlackNode) {
 }
 
 // creates a new node with specified value and default fields.
-func (t *redBlackTree) spawn(rating int, playerId string) *redBlackNode {
+func (t *redBlackTree) spawn(rating float64, playerId string) *redBlackNode {
 	return &redBlackNode{
 		key:    nodeKey{rating: rating, playerId: playerId},
 		isRed:  true,
